Passed the Test Driver. Finishing up my own tests

+++ src/main/java/Model/Event.java
@@ -14,10 +14,10 @@ public class Event {
     private String personID;
 
     /** Latitude of the event’s location */
-    private double latitude;
+    private float latitude;
 
     /** Longitude of the event’s location */
-    private double longitude;
+    private float longitude;
 
     /** Name of country where event occurred */
     private String country;
@@ -50,7 +50,7 @@ public class Event {
      * @param eventType The type of event (i.e. birth, death, baptism, wedding, etc.)
      * @param year Year event occurred
      * */
-    public Event(String eventID, String associatedUsername, String personID, double latitude, double longitude, String country, String city, String eventType, int year) {
+    public Event(String eventID, String associatedUsername, String personID, float latitude, float longitude, String country, String city, String eventType, int year) {
         this.eventID = eventID;
         this.associatedUsername = associatedUsername;
         this.personID = personID;
@@ -86,19 +86,19 @@ public class Event {
         this.personID = personID;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
